# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EarnQA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /referrals/click:
    post:
      operationId: logReferralClick
      tags: [referrals]
      summary: Log a referral link click
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralClickBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
  /referrals/claim:
    post:
      operationId: claimReferral
      tags: [referrals]
      summary: Claim a referral code after signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralClaimBody"
      responses:
        "200":
          description: Claim result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralClaimResponse"
...
  /referrals/admin/{id}/status:
    patch:
      operationId: patchAdminReferralStatus
      tags: [referrals]
      summary: Admin — update a referral's status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralStatusPatchBody"
      responses:
        "200":
          description: Updated referral
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralEntry"
  /referrals/admin/{id}/reverse:
    post:
      operationId: reverseAdminReferral
      tags: [referrals]
      summary: Admin — reverse a referral and reclaim bonuses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reversal result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  reversedCents:
                    type: number
                required:
                  - ok
                  - reversedCents
components:
  schemas:
...
    ReferralEntry:
      type: object
      properties:
        id:
          type: integer
        referrerUserId:
          type: string
        referredUserId:
          type: string
        referralCodeUsed:
          type: string
        signupBonusCents:
          type: number
        answerBonusCentsTotal:
          type: number
        signupBonusGrantedAt:
          type: ["string", "null"]
        referrerClickIp:
          type: ["string", "null"]
        referredSignupIp:
          type: ["string", "null"]
        referredUserAgent:
          type: ["string", "null"]
        status:
          type: string
        fraudFlags:
          type: array
          items:
            type: string
        referrerName:
          type: string
        referrerEmail:
          type: string
        referredName:
          type: string
        referredEmail:
          type: string
        referredUserName:
          type: string
        referredUserEmail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - referrerUserId
        - referredUserId
        - referralCodeUsed
        - signupBonusCents
        - answerBonusCentsTotal
        - status
        - createdAt
        - updatedAt
...
    ReferralClickBody:
      type: object
      properties:
        referralCode:
          type: string
      required:
        - referralCode
    ReferralClaimBody:
      type: object
      properties:
        referralCode:
          type: string
      required:
        - referralCode
    ReferralClaimResponse:
      type: object
      properties:
        ok:
          type: boolean
        alreadyClaimed:
          type: boolean
        status:
          type: string
        fraudFlags:
          type: array
          items:
            type: string
      required:
        - ok
...
    ReferralStatusPatchBody:
      type: object
      properties:
        status:
          type: string
          enum: [approved, flagged, rejected, pending]
      required:
        - status